fix(show-hide-input): fix binding id attribute
This fixes a bug when ShowHideInputDirective tries to find element id when using binding. Accessing the element id through nativeElement works only if the id is statically assigned `id="test"`. This enables the directive to access the element id in case of using angular binding `[id]="someValue"`. The directive no longer access the id through the native element but instead from an `@input` prop.